AiInternTask/chatbot_theme_identifier/backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyPDF2
pytesseract
Pillow
requests
python-dotenv
python-docx
pinecone
fastembed